+++ CHANGELOG.md
@@ -6,6 +6,9 @@
 
 ### Features
 
+- Introducing an example for proving knowledge of exponent
+- Add api to get SRS size.
+
 ### Improvements
 
 - Derive `Debug`, `Snafu` on `enum TaggedBlobError`